+++ app/src/app/ui/ui.module.js
@@ -60,6 +60,11 @@ angular.module('triAngularUI', ['ngCookies', 'webfont-loader'])
             }
         }
     })
+    .state('admin-panel.default.ui-weather-icons', {
+        url: '/ui/weather-icons',
+        controller: 'WeatherIconsController',
+        templateUrl: 'app/ui/weather-icons.tmpl.html'
+    })
     .state('admin-panel.default.ui-fa-icons', {
         url: '/ui/fa-icons',
         controller: 'FaIconsController',
@@ -156,6 +161,10 @@ angular.module('triAngularUI', ['ngCookies', 'webfont-loader'])
         },{
             name: 'MENU.UI.TYPOGRAPHY',
             state: 'admin-panel.default.ui-typography',
+            type: 'link',
+        },{
+            name: 'MENU.UI.WEATHER_ICONS',
+            state: 'admin-panel.default.ui-weather-icons',
             type: 'link',
         }]
     });